### Description

This PR incorporates modifications to the block validator within the op-node, which is utilized by both the sender (sequencer) and receiver. The block expiration check has been increased from 60 seconds to 20 minutes. This change was prompted by scenarios where the sequencer could not maintain a block production rate of one block per second. Under the previous validation logic, if blocks were produced slowly for a period and the anticipated block time was more than 60 seconds apart from the current time, the sequencer would fail the check and be unable to broadcast the block, preventing other nodes from syncing unsafe blocks. With this change, the time has been increased to 20 minutes. Please note that the 20-minute duration is not the time we maintain synchronization, but a relative time difference. (For example, if the sequencer’s block production speed drops to half, at 2 seconds per block, then the actual duration for which the sequencer continues broadcasting blocks would be approximately 40 minutes.)

### Rationale
Two adjustments have been made:
1.The time duration within the block validator has been extended from 60s to 20min.
2.Enhancements were made to the seenMessagesTTL and blockHeightLRU caches to guard against potential stale block attacks that could arise due to the validator changes. The seenMessagesTTL serves as a preliminary logic cache before entering the validator, where a block hitting this cache would not proceed to further validation and processing. The blockHeightLRU cache is used to keep track of the number of validations for duplicate blocks to signal other nodes with an 'ignore' or 'reject'. As our validation effective time has increased from 1 minute to 20 minutes, the corresponding caches also required expansion to prevent possible block broadcast attacks.

### Example

N/A

### Changes

Notable changes:
* BuildBlocksValidator config change
* seenMessagesTTL and blockHeightLRU extended